requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
pyodbc>=5.0.1
python-dotenv>=1.0.0
hypercorn>=0.15.0
asyncio>=3.4.3
openai>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
python-dateutil>=2.8.2
